shift homecommunity creation for tests in seed directory-tree
