+++ backend/src/graphql/resolver/EmailOptinCodes.test.ts
@@ -8,10 +8,9 @@ import { GraphQLError } from 'graphql'
 import { testEnvironment, cleanDB } from '@test/helpers'
 
 import { CONFIG } from '@/config'
+import { writeHomeCommunityEntry } from '@/seeds/community'
 import { createUser, setPassword, forgotPassword } from '@/seeds/graphql/mutations'
 import { queryOptIn } from '@/seeds/graphql/queries'
-
-import { createHomeCommunity } from './util/communities'
 
 let mutate: ApolloServerTestClient['mutate'],
   query: ApolloServerTestClient['query'],
@@ -48,7 +47,7 @@ describe('EmailOptinCodes', () => {
       lastName: 'Lustig',
       language: 'de',
     }
-    await createHomeCommunity()
+    await writeHomeCommunityEntry()
     const {
       data: { createUser: user },
     } = await mutate({ mutation: createUser, variables })
